trend_classifier/pyproject.toml
[tool.poetry]
name = "trend-classifier"
version = "0.1.0"
description = ""
authors = ["MaddoSaru <riquelme.madmonkey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.3.3"
snowflake-connector-python = "^4.0.0"
python-dotenv = "^1.2.1"
black = "^25.11.0"
prophet = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
